updated the fragment that hosts the recycler view
and started to update user stories and documentation for the project
