+++ docs/index.md
@@ -6,6 +6,12 @@ Keep in Touch is an app focused on helping young people who forget to respond to
 
 KiT accomplishes this by using text message metadata from inbound and outbound messages and corresponding phone contacts. It utilizes the built in messages app in Android to send short  _keep in touch_ messages to friends and family. When a user first downloads the app and signs in via Google, they can select a preferred user-type based on user demographic information. Once completed each user will have access to smart auto-replies to your contacts, and a more catered experience.
 
+KiT intends to make it easier to reply to contacts, who've already messaged you. The purpose of KiT is to make responding back effortless.
+The implications of what this app allows are interesting, because it can change how people communicate through SMS in the real world. During the Covid-19 Pandemic,
+we've all experienced the way that communication suffers. KiT aims to alleviate the natural issue that arise from SMS communication, continuity. Currently, most native
+text messaging apps don't have a reminder feature built in. Even if they did, users may be reluctant to send a message unless it's effortless. Lately, services like Gmail
+have offered similar functionality, but there's no app that reminds different user groups for SMS messaging.  
+
 ### Key features
 
 * Ignore contact
