+++ docs/work/intendedUsers.md
@@ -2,9 +2,8 @@
 
 #### Teens![Teens](../img/teens.jpg)
 
-> As a teen, I want to make sure I don't forget to text my friends back so that  I can stay connected with what's going on and foster my friendships.
-
-> As a teen with anxiety, I don't want to miss out. Kit will help me step past the "over-thinking" in my head, by nudging me with reminders. So that I can stay connected digitally and let my friends know how I'm doing. That way they can check up on me in real life.
+> As a teen, I want to make use of KiT's auto-replies based on user type, so I can respond in a fun way. The reminder durations will help to make sure I respond to my friends. That way I can stay connected with what's going on and foster my friendships.
+I know KiT will help me step past the "over-thinking" in my head, by nudging me with reminders. So that I can stay connected digitally and let my friends know how I'm doing. That way they can check up on me in real life.
 
 #### Millennials
 
@@ -12,9 +11,9 @@
 --- | ---
 <img src="../img/millenials.jpg" width="400px"/> | <img src="../img/millenials2.jpg" width="400px" />
 
-> As a young professional, I want to respond to my friends, acquaintances, and work connections in a timely manner so I can build my social network and create deeper, more meaningful connections with the people I've met.
+> As a young professional, I want to use KiT to respond in a timely-manner to my friends, acquaintances, and work connections. I will use the reminder duration to keep messaging contacts back, so that I can build my social network and create deeper, more meaningful connections with the people I've met.
 
-> As a recent college graduate, I can barely handle \#adulting. I want an easy way to review all my old messages in a glance, to make sure I didn't forget to message my friends back. That way I can stay on top of my messages while I learn how to manage paying bills, cooking, and \#adulting. 
+> As a recent college graduate, I can barely handle \#adulting. I want an easy way to review all my old messages at a glance, to make sure I didn't forget to message my friends back. I can use KiT's reminder duration to  That way I can stay on top of my messages while I learn how to manage paying bills, cooking, and \#adulting. 
 
 #### Parents ![Parents](../img/parents2.jpg)
